feat: add new optional org parameter to newRepo function
With this new parameter, we can use settings from the org in the repo settings, and adapt some of the default repo settings depending on the organization settings. 

We do this for actions_can_approve_pull_request_reviews, that takes the same value of the organization settings if the org is given. Otherwise, we just use the same default value as before.